Install BP during Behat tests.
This setup works by copying the content of the `BP_SRC_DIR` to the
WordPress install that is installed by the wp-cli utilities. BP must
then be installed and tested components activated.

When the commands are merged into BuddyPress, we can set `BP_SRC_DIR`
automatically. Until then, it should be set as an environment variable.

This changeset also includes fixes for the xprofile-group tests to
demonstrate how the new `Given a BP install` condition works, and to
show how the tests should be organized to minimize duplication. See
the `wp user` tests for an example.

See #8.